+++ playground/nuxt.config.ts
@@ -22,13 +22,13 @@ export default <NuxtConfig> {
       'https://upload.wikimedia.org'
     ],
     twicpics: {
-      baseURL: 'https://nuxt-demo.twic.pics'
+      baseURL: 'https://demo.twic.pics/'
     },
     storyblok: {
-      baseURL: 'https://img2.storyblok.com'
+      baseURL: 'https://img2.storyblok.com/'
     },
     cloudinary: {
-      baseURL: 'https://res.cloudinary.com/nuxt/image/upload'
+      baseURL: 'https://res.cloudinary.com/demo/image/upload'
     },
     fastly: {
       baseURL: 'https://www.fastly.io'
@@ -41,6 +41,9 @@ export default <NuxtConfig> {
     },
     sanity: {
       projectId: 'j1o4tmjp'
+    },
+    vercel: {
+      baseURL: 'https://image-component.nextjs.gallery/_next/image'
     },
     providers: {
       custom: {
@@ -57,9 +60,6 @@ export default <NuxtConfig> {
           height: 50
         }
       }
-    },
-    intersectOptions: {
-      rootMargin: '50px'
     }
   }
 }