+++ src/module.ts
@@ -1,16 +1,20 @@
-import { resolve } from 'path'
+import { resolve } from 'upath'
+
+import type { Module } from '@nuxt/types'
 import defu from 'defu'
+
+import { setupStaticGeneration } from './generate'
+import { createIPXMiddleware } from './ipx'
+import { resolveProviders, detectProvider } from './provider'
 import type { ModuleOptions, CreateImageOptions } from './types'
 import { pick, pkg } from './utils'
-import { setupStaticGeneration } from './generate'
-import { resolveProviders } from './provider'
-import { createIPXMiddleware } from './ipx'
 
-async function imageModule (moduleOptions: ModuleOptions) {
+const imageModule: Module<ModuleOptions> = async function imageModule (moduleOptions) {
   const { nuxt, addPlugin, addServerMiddleware } = this
 
   const defaults: ModuleOptions = {
-    provider: 'static',
+    staticFilename: '[publicPath]/image/[hash][ext]',
+    provider: 'auto',
     presets: {},
     dir: resolve(nuxt.options.srcDir, nuxt.options.dir.static),
     domains: [],
@@ -27,24 +31,32 @@ async function imageModule (moduleOptions: ModuleOptions) {
     },
     internalUrl: '',
     providers: {},
-    static: {},
-    intersectOptions: {}
+    static: {}
   }
 
   const options: ModuleOptions = defu(moduleOptions, nuxt.options.image, defaults)
 
-  options.provider = process.env.NUXT_IMAGE_PROVIDER || options.provider || 'static'
+  options.provider = detectProvider(options.provider)
+
+  options[options.provider] = options[options.provider] || {}
 
   const imageOptions: Omit<CreateImageOptions, 'providers'> = pick(options, [
     'screens',
     'presets',
-    'provider',
-    'intersectOptions'
+    'provider'
   ])
 
+  options.static = options.static || {}
   options.static.domains = options.domains
 
-  const providers = await resolveProviders(nuxt, options)
+  const providers = resolveProviders(nuxt, options)
+
+  // Run setup
+  for (const p of providers) {
+    if (typeof p.setup === 'function') {
+      await p.setup(p, options, nuxt)
+    }
+  }
 
   // Transpile and alias runtime
   const runtimeDir = resolve(__dirname, 'runtime')
@@ -70,25 +82,26 @@ async function imageModule (moduleOptions: ModuleOptions) {
     })
   })
 
+  // transform asset urls that pass to `src` attribute on image components
   nuxt.options.build.loaders = defu({
-    vue: { transformAssetUrls: { 'nuxt-img': 'src', 'nuxt-picture': 'src' } }
+    vue: { transformAssetUrls: { 'nuxt-img': 'src', 'nuxt-picture': 'src', NuxtPicture: 'src', NuxtImg: 'src' } }
   }, nuxt.options.build.loaders || {})
 
   nuxt.hook('generate:before', () => {
     setupStaticGeneration(nuxt, options)
   })
 
-  const LruCache = require('lru-cache')
+  const LruCache = await import('lru-cache').then(r => r.default || r)
   const cache = new LruCache()
-  nuxt.hook('vue-renderer:context', (ssrContext) => {
+  nuxt.hook('vue-renderer:context', (ssrContext: any) => {
     ssrContext.cache = cache
   })
 
-  nuxt.hook('listen', (_, listener) => {
+  nuxt.hook('listen', (_: any, listener: any) => {
     options.internalUrl = `http://localhost:${listener.port}`
   })
 }
 
-(imageModule as any).meta = pkg
+  ; (imageModule as any).meta = pkg
 
 export default imageModule