+++ src/runtime/components/nuxt-img.vue
@@ -7,17 +7,25 @@
 </template>
 
 <script lang="ts">
+import type { DefineComponentWithMixin } from '../../types/vue'
+import type { ImageSizes } from '../../types'
 import { imageMixin } from './image.mixin'
-import { EMPTY_GIF, lazyMixin } from './lazy.mixin'
 
 import { parseSize } from '~image'
 
-export default {
+const defineComponent: DefineComponentWithMixin = (opts: any) => opts
+
+type NAttrs = typeof imageMixin['nImgAttrs'] & {
+    sizes?: string
+    srcset?: string
+}
+
+export default defineComponent({
   name: 'NuxtImg',
-  mixins: [imageMixin, lazyMixin],
+  mixins: [imageMixin],
   computed: {
-    nAttrs () {
-      const attrs: any = this.nImgAttrs
+    nAttrs (): NAttrs {
+      const attrs: NAttrs = this.nImgAttrs
       if (this.sizes) {
         const { sizes, srcset } = this.nSizes
         attrs.sizes = sizes
@@ -25,16 +33,11 @@ export default {
       }
       return attrs
     },
-    nSrc () {
-      if (this.lazyLoad) {
-        return EMPTY_GIF
-      }
-      if (this.sizes) {
-        return this.nSizes.src
-      }
-      return this.$img(this.src, this.nModifiers, this.nOptions)
+    nSrc (): string {
+      return this.sizes ? this.nSizes.src : this.$img(this.src, this.nModifiers, this.nOptions)
     },
-    nSizes () {
+    /* eslint-disable no-undef */
+    nSizes (): ImageSizes {
       return this.$img.getSizes(this.src, {
         ...this.nOptions,
         sizes: this.sizes,
@@ -55,5 +58,5 @@ export default {
       }
     }
   }
-}
+})
 </script>