+++ src/types/image.ts
@@ -34,7 +34,6 @@ export interface CreateImageOptions {
   }
   presets: { [name: string]: ImageOptions }
   provider: string
-  intersectOptions: object
   screens?: Record<string, number>,
 }
 
@@ -51,13 +50,22 @@ export interface ResolvedImage {
   getMeta?: () => Promise<ImageInfo>
 }
 
-export interface $Img {
+export interface ImageSizes {
+  srcset: string
+  sizes: string
+  src: string
+}
+
+export interface Img {
   (source: string, modifiers?: ImageOptions['modifiers'], options?: ImageOptions): ResolvedImage['url']
   options: CreateImageOptions
   getImage: (source: string, options?: ImageOptions) => ResolvedImage
-  getSizes: (source: string, options?: ImageOptions, sizes?: string[]) => { srcset: string, sizes: string }
+  getSizes: (source: string, options?: ImageOptions, sizes?: string[]) => ImageSizes
   getMeta: (source: string, options?: ImageOptions) => Promise<ImageInfo>
-  [preset: string]: $Img['options'] | $Img['getImage'] | $Img['getSizes'] | $Img['getMeta'] | $Img /* preset */
+}
+
+export type $Img = Img & {
+  [preset: string]: $Img
 }
 
 export interface ImageCTX {