+++ src/types/module.ts
@@ -1,10 +1,15 @@
 import type { IPXOptions } from 'ipx'
 import type { ImageOptions, CreateImageOptions } from './image'
 
-export interface InputProvider<T=any> {
+// eslint-disable-next-line no-use-before-define
+export type ProviderSetup = (providerOptions: ImageModuleProvider, moduleOptions: ModuleOptions, nuxt: any)
+  => void | Promise<void>
+
+export interface InputProvider<T = any> {
   name?: string
   provider?: string
   options?: T
+  setup?: ProviderSetup
 }
 
 export interface ImageProviders {
@@ -20,6 +25,7 @@ export interface ImageProviders {
 
 // TODO: use types from CreateImageOptions
 export interface ModuleOptions extends ImageProviders {
+  staticFilename: string,
   provider: CreateImageOptions['provider']
   presets: { [name: string]: ImageOptions }
   dir: string
@@ -27,6 +33,16 @@ export interface ModuleOptions extends ImageProviders {
   sharp: {}
   screens: CreateImageOptions['screens'],
   internalUrl: string
-  intersectOptions: CreateImageOptions['intersectOptions']
   providers: { [name: string]: InputProvider | any } & ImageProviders
+  [key: string]: any
+}
+
+export interface ImageModuleProvider {
+  name: string
+  importName: string
+  options: any
+  provider: string
+  runtime: string
+  runtimeOptions: any
+  setup: ProviderSetup
 }