+++ test/fixture/base/providers/random/runtime.ts
@@ -1,8 +1,0 @@
-
-export default {
-  getImage () {
-    return {
-      url: 'https://source.unsplash.com/random/600x400'
-    }
-  }
-}