+++ test/fixture/providers/random/index.ts
@@ -1,0 +1,5 @@
+export function getImage () {
+  return {
+    url: 'https://source.unsplash.com/random/600x400'
+  }
+}