+++ test/setup.ts
@@ -1,0 +1,4 @@
+import consola from 'consola'
+
+consola.wrapAll()
+consola.mockTypes(() => jest.fn())