+++ test/unit/utils/tick.ts
@@ -1,0 +1,7 @@
+export async function nextTick () {
+  await new Promise(resolve =>
+    process.nextTick(() =>
+      resolve(null)
+    )
+  )
+}